requirements.txt
anthropic>=0.40.0
python-docx>=1.1.0
lxml>=5.0.0
docx2pdf>=0.1.8
